requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
PyYAML>=6.0.2
uritemplate>=4.1.1
psycopg2-binary>=2.9.9  
django-cors-headers>=4.3.1
python-decouple>=3.8       
gunicorn>=22.0.0           
whitenoise>=6.6.0          
black>=24.8.0              
isort>=5.13.2             
python-dotenv>=1.0.0     
mysqlclient>=2.2.4
dotenv>=0.0.5
pymysql>=1.0.3